chore: Move "Delete Track" function over
